+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the `dom_query` crate will be documented in this file.
 
+## [Unreleased]
+### Changed
+- Revised `NodeRef::find_descendants` (requires `mini_selector` feature). This method now supports `Adjacent (+)` and `Sibling (~)` combinators.
+
 ## [0.22.0] - 2025-09-03
 
 ### Added